feat: add camera_web barrel file
