+++ packages/camera/camera_web/lib/camera_web.dart
@@ -1,193 +1,3 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
+library camera_web;
 
-import 'dart:async';
-import 'dart:html' as html;
-import 'dart:math';
-
-import 'package:camera_platform_interface/camera_platform_interface.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_web_plugins/flutter_web_plugins.dart';
-
-/// The web implementation of [CameraPlatform].
-///
-/// This class implements the `package:camera` functionality for the web.
-class CameraPlugin extends CameraPlatform {
-  /// Registers this class as the default instance of [CameraPlatform].
-  static void registerWith(Registrar registrar) {
-    CameraPlatform.instance = CameraPlugin();
-  }
-
-  /// The current browser window used to access device cameras.
-  @visibleForTesting
-  html.Window? window;
-
-  @override
-  Future<List<CameraDescription>> availableCameras() {
-    throw UnimplementedError('availableCameras() is not implemented.');
-  }
-
-  @override
-  Future<int> createCamera(
-    CameraDescription cameraDescription,
-    ResolutionPreset? resolutionPreset, {
-    bool enableAudio = false,
-  }) {
-    throw UnimplementedError('createCamera() is not implemented.');
-  }
-
-  @override
-  Future<void> initializeCamera(
-    int cameraId, {
-    ImageFormatGroup imageFormatGroup = ImageFormatGroup.unknown,
-  }) {
-    throw UnimplementedError('initializeCamera() is not implemented.');
-  }
-
-  @override
-  Stream<CameraInitializedEvent> onCameraInitialized(int cameraId) {
-    throw UnimplementedError('onCameraInitialized() is not implemented.');
-  }
-
-  @override
-  Stream<CameraResolutionChangedEvent> onCameraResolutionChanged(int cameraId) {
-    throw UnimplementedError('onCameraResolutionChanged() is not implemented.');
-  }
-
-  @override
-  Stream<CameraClosingEvent> onCameraClosing(int cameraId) {
-    throw UnimplementedError('onCameraClosing() is not implemented.');
-  }
-
-  @override
-  Stream<CameraErrorEvent> onCameraError(int cameraId) {
-    throw UnimplementedError('onCameraError() is not implemented.');
-  }
-
-  @override
-  Stream<VideoRecordedEvent> onVideoRecordedEvent(int cameraId) {
-    throw UnimplementedError('onVideoRecordedEvent() is not implemented.');
-  }
-
-  @override
-  Stream<DeviceOrientationChangedEvent> onDeviceOrientationChanged() {
-    throw UnimplementedError(
-      'onDeviceOrientationChanged() is not implemented.',
-    );
-  }
-
-  @override
-  Future<void> lockCaptureOrientation(
-    int cameraId,
-    DeviceOrientation orientation,
-  ) {
-    throw UnimplementedError('lockCaptureOrientation() is not implemented.');
-  }
-
-  @override
-  Future<void> unlockCaptureOrientation(int cameraId) {
-    throw UnimplementedError('unlockCaptureOrientation() is not implemented.');
-  }
-
-  @override
-  Future<XFile> takePicture(int cameraId) {
-    throw UnimplementedError('takePicture() is not implemented.');
-  }
-
-  @override
-  Future<void> prepareForVideoRecording() {
-    throw UnimplementedError('prepareForVideoRecording() is not implemented.');
-  }
-
-  @override
-  Future<void> startVideoRecording(int cameraId, {Duration? maxVideoDuration}) {
-    throw UnimplementedError('startVideoRecording() is not implemented.');
-  }
-
-  @override
-  Future<XFile> stopVideoRecording(int cameraId) {
-    throw UnimplementedError('stopVideoRecording() is not implemented.');
-  }
-
-  @override
-  Future<void> pauseVideoRecording(int cameraId) {
-    throw UnimplementedError('pauseVideoRecording() is not implemented.');
-  }
-
-  @override
-  Future<void> resumeVideoRecording(int cameraId) {
-    throw UnimplementedError('resumeVideoRecording() is not implemented.');
-  }
-
-  @override
-  Future<void> setFlashMode(int cameraId, FlashMode mode) {
-    throw UnimplementedError('setFlashMode() is not implemented.');
-  }
-
-  @override
-  Future<void> setExposureMode(int cameraId, ExposureMode mode) {
-    throw UnimplementedError('setExposureMode() is not implemented.');
-  }
-
-  @override
-  Future<void> setExposurePoint(int cameraId, Point<double>? point) {
-    throw UnimplementedError('setExposurePoint() is not implemented.');
-  }
-
-  @override
-  Future<double> getMinExposureOffset(int cameraId) {
-    throw UnimplementedError('getMinExposureOffset() is not implemented.');
-  }
-
-  @override
-  Future<double> getMaxExposureOffset(int cameraId) {
-    throw UnimplementedError('getMaxExposureOffset() is not implemented.');
-  }
-
-  @override
-  Future<double> getExposureOffsetStepSize(int cameraId) {
-    throw UnimplementedError('getExposureOffsetStepSize() is not implemented.');
-  }
-
-  @override
-  Future<double> setExposureOffset(int cameraId, double offset) {
-    throw UnimplementedError('setExposureOffset() is not implemented.');
-  }
-
-  @override
-  Future<void> setFocusMode(int cameraId, FocusMode mode) {
-    throw UnimplementedError('setFocusMode() is not implemented.');
-  }
-
-  @override
-  Future<void> setFocusPoint(int cameraId, Point<double>? point) {
-    throw UnimplementedError('setFocusPoint() is not implemented.');
-  }
-
-  @override
-  Future<double> getMaxZoomLevel(int cameraId) {
-    throw UnimplementedError('getMaxZoomLevel() is not implemented.');
-  }
-
-  @override
-  Future<double> getMinZoomLevel(int cameraId) {
-    throw UnimplementedError('getMinZoomLevel() is not implemented.');
-  }
-
-  @override
-  Future<void> setZoomLevel(int cameraId, double zoom) {
-    throw UnimplementedError('setZoomLevel() is not implemented.');
-  }
-
-  @override
-  Widget buildPreview(int cameraId) {
-    throw UnimplementedError('buildPreview() is not implemented.');
-  }
-
-  @override
-  Future<void> dispose(int cameraId) {
-    throw UnimplementedError('dispose() is not implemented.');
-  }
-}
+export 'src/camera_web.dart';
